backend/AI/Matching/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
python-multipart
sqlalchemy
pymysql
python-dotenv
cryptography